Move cmd function impl from gost-core to engine-related translation unit
