+++ CMakeLists.txt
@@ -214,6 +214,7 @@ set(GOST_ENGINE_SOURCE_FILES
         gost_eng_digest.c
         gost_eng_digest_define.c
         gost_eng_cipher.c
+        gost_eng_cmd.c
         gost_cipher_ctx_evp.c
         gost_eng_pmeth.c
         )
