+++ gost_ctl.c
@@ -10,7 +10,6 @@
 #include <string.h>
 #include <openssl/crypto.h>
 #include <openssl/err.h>
-#include <openssl/engine.h>
 #include <openssl/buffer.h>
 #include "gost_lcl.h"
 
@@ -29,16 +28,6 @@ void gost_param_free()
 
 }
 
-int gost_control_func(ENGINE *e, int cmd, long i, void *p, void (*f) (void))
-{
-    int param = cmd - ENGINE_CMD_BASE;
-    int ret = 0;
-    if (param < 0 || param > GOST_PARAM_MAX) {
-        return -1;
-    }
-    ret = gost_set_default_param(param, p);
-    return ret;
-}
 
 const char *get_gost_engine_param(int param)
 {
