+++ test_ecdhe.c
@@ -1,5 +1,6 @@
 #include <openssl/ec.h>
 #include <openssl/bn.h>
+#include <openssl/err.h>
 #include <stdio.h>
 #include <string.h>
 #include <stdlib.h>
